TVG-102 Blueprint for SearchSubscription endpoints
