+++ api.py
@@ -13,12 +13,12 @@ from api_blueprints import (
     reminder_blueprint,
     reminders_blueprint,
     search_items_blueprint,
+    search_subscription_blueprint,
     shows_blueprint,
     show_episodes_blueprint,
 )
 from database import engine
-from database.models import SearchItem, User, UserSearchSubscription
-from exceptions.DatabaseError import InvalidSubscriptions
+from database.models import User
 
 app = Flask(__name__, template_folder='frontend/build', static_folder='frontend/build/assets')
 CORS(app)
@@ -78,6 +78,10 @@ app.register_blueprint(guide_blueprint, url_prefix="/api/guide")
 app.register_blueprint(reminder_blueprint, url_prefix="/api/reminder")
 app.register_blueprint(reminders_blueprint, url_prefix="/api/reminders")
 app.register_blueprint(search_items_blueprint, url_prefix="/api/search-item")
+app.register_blueprint(
+    search_subscription_blueprint,
+    url_prefix="/api/users/<string:username>/subscriptions"
+)
 app.register_blueprint(shows_blueprint, url_prefix="/api/shows")
 app.register_blueprint(show_episodes_blueprint, url_prefix="/api/show-episode") 
 
@@ -90,54 +94,6 @@ def get_user(username: str):
     if user:
         return user.to_dict()
     return {'message': f'An account with the username {username} could not be found'}, 404
-
-@app.route('/api/users/<string:username>/subscriptions', methods=['GET'])
-def get_user_subscriptions(username: str):
-    session = Session(engine)
-    user = User.search_for_user(username, session)
-    if user:
-        viewed_user = User.search_for_user(username, session)
-        user_subscriptions = UserSearchSubscription.get_user_subscriptions(session, viewed_user.id)
-        return [subscription.to_dict() for subscription in user_subscriptions]           
-    return {'message': f'A user with the username {username} could not be found'}, 404
-
-@app.route('/api/users/<string:username>/subscriptions', methods=['POST'])
-@jwt_required()
-def add_user_subscriptions(username: str):
-    session = Session(engine)
-    user = User.search_for_user(username, session)
-    if user:
-        current_user: User = get_current_user()
-        if current_user.username != username:
-            return {'message': "You are not able to update this user's details"}, 403
-        body: list[str] = request.json
-        try:
-            user_subscriptions: list[UserSearchSubscription] = []
-            for show in body:
-                search_item = SearchItem.get_search_item(show, session)
-                user_subscriptions.append(UserSearchSubscription(user.id, search_item.id))
-            UserSearchSubscription.add_subscription_list(user_subscriptions, session)
-            return user.to_dict()
-        except InvalidSubscriptions as err:
-            return {'message': str(err)}, 400            
-    return {'message': f'A user with the username {username} could not be found'}, 404
-
-@app.route('/api/users/subscriptions/<string:subscription_id>', methods=['DELETE'])
-@jwt_required()
-def delete_user_subscription(subscription_id: str):
-    session = Session(engine)
-    subscription = UserSearchSubscription.get_subscription_by_id(subscription_id, session)
-    if not subscription:
-        return { 'message': f'This subscription could not be found' }, 404
-    if not subscription.user:
-        return { 'message': f'Unable to find the user account' }, 400
-    if not subscription.search_item:
-        return { 'message': f'This search item does not exist' }, 400
-    try:
-        subscription.remove_subscription(session)
-        return "", 204
-    except InvalidSubscriptions as err:
-        return {'message': str(err)}, 400 
 
 @app.route('/api/user/<string:username>/promote', methods=['PATCH'])
 @jwt_required()
