+++ api_blueprints/__init__.py
@@ -1,5 +1,6 @@
 from api_blueprints.guide import guide_blueprint
 from api_blueprints.reminders import reminder_blueprint, reminders_blueprint
 from api_blueprints.search_items import search_items_blueprint
+from api_blueprints.search_subscriptions import search_subscription_blueprint
 from api_blueprints.shows import shows_blueprint
 from api_blueprints.show_episodes import show_episodes_blueprint
